pyproject.toml
[project]
name = "pubsub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-messaging-webpubsubservice>=1.2.1",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.9",
]
